preversion 1.13.0c01

+++ src/.NET Core 3.1/Inertia/Network/Common/Protocol/NetworkMessage.cs
@@ -4,7 +4,7 @@ namespace Inertia.Network
 {
     public abstract class NetworkMessage : IAutoSerializable, ISerializableObject
     {
-        public abstract uint MessageId { get; }
+        public abstract ushort MessageId { get; }
 
         public virtual bool UseAutoSerialization { get; } = true;
